best render of gender data, cleaned up components, commented out code, console.logs, edited CSS
