+++ client/components/HomePage/Components/StateMap.js
@@ -1,6 +1,6 @@
 import React, {useRef, useEffect} from 'react'
 import {geoAlbersUsa, geoPath, select, scaleSqrt, easeLinear, max} from 'd3'
-import usData from './usData.json'
+import usData from '../Assets/usData.json'
 
 export const StateMap = data => {
   // const [geographies, setGeographies] = useState([])
@@ -28,8 +28,8 @@ export const StateMap = data => {
         // Run data through pathGenerator
         .attr('d', pathGenerator)
         // Style
-        .style('fill', 'rgba(38,50,56)')
-        .style('stroke', '#011627')
+        .style('fill', '#384F56')
+        .style('stroke', '#080B0C')
         .style('stroke-width', 2)
       const maxCases = max(data.data.map(d => d.positive))
 
@@ -48,7 +48,9 @@ export const StateMap = data => {
         // Add class for reference
         .attr('class', 'circle')
         // style
-        .style('fill', '#E91E')
+
+        //.style('fill', '#E91E')
+        .style('fill', '#80ed99')
         .style('fill-opacity', 0.7)
         // Update x-position
         .attr('cx', d => projection([d.longitude, d.latitude])[0])
