+++ client/components/HomePage/Components/navbar.js
@@ -3,8 +3,9 @@ import React from 'react'
 import {Link} from 'react-router-dom'
 
 export const Navbar = () => (
-  <div>
+  <div className="flex">
     <h1>coviz</h1>
+
     <nav>
       <div>
         {/* The navbar will show these links before you log in */}
@@ -14,6 +15,7 @@ export const Navbar = () => (
         <Link to="/about">About</Link>
       </div>
     </nav>
-    <hr />
+
+    {/* <hr /> */}
   </div>
 )
