+++ client/components/index.js
@@ -3,9 +3,9 @@
  * This way, we can import all of our components from the same place, rather than
  * having to figure out which file they belong to!
  */
-export {App} from './homePage/App'
-export {Chart} from './chartd3/Chart'
-export {Navbar} from './navbar'
-export {EthnicityChart} from './homePage/ethnicity'
-export {About} from './About'
-export {GenderChart} from './homePage/GenderChart'
+export {App} from './HomePage/Components/App'
+export {Chart} from './D3Charts/Chart'
+export {Navbar} from './HomePage/Components/navbar'
+export {EthnicityChart} from './HomePage/Components/EthnicityChart'
+export {About} from './HomePage/Components/About'
+export {GenderChart} from './HomePage/Components/GenderChart'
