+++ client/store/usDataByDate.js
@@ -13,20 +13,7 @@ export const fetchAllDateData = data => ({
 export const fetchAllDateDataThunk = () => {
   return async dispatch => {
     try {
-      // const array = date.split('/')
-      // const year = array.pop()
-      // array.unshift(year)
-      // if (array[1].length <= 1) {
-      //   let temp = array[1]
-      //   array[1] = `0${temp}`
-      // }
-      // if (array[2].length <= 1) {
-      //   let temp = array[2]
-      //   array[2] = `0${temp}`
-      // }
-      // const newDate = array.join('')
       const {data} = await axios.get(`api/covidDaily`)
-      // console.log('thunk data', data)
       dispatch(fetchAllDateData(data))
     } catch (err) {
       console.error(err)
