+++ client/store/usDataByDate.spec.js
@@ -34,7 +34,7 @@ describe('usDataByDate redux', () => {
   })
   it('fetchAllDateData action creator returns a valid action', () => {
     const example = fetchAllDateData(data)
-    console.log('EXAMPLE', example)
+
     expect(fetchAllDateData(data)).to.deep.equal({
       type: 'GET_ALL_DATE_DATA',
       data
