--- script/covidrace(no).csv
@@ -1,51 +1,0 @@
-State,Non-Hispanic White,Non-Hispanic Black or African American,Non-Hispanic American Indian or Alaska Native,Non-Hispanic Asian,Non-Hispanic Native Hawaiian or Other Pacific Islander,Hispanic or Latino,Other
-United States,138695,46756,2689,9448,459,49339,2184
-Alabama,2815,1497,0,22,0,105,0
-Arizona,2612,199,779,95,16,1932,140
-Arkansas,1867,375,13,24,53,117,0
-California,5673,1323,72,2164,90,9039,144
-Colorado,1670,163,34,99,13,707,22
-Connecticut,3370,703,0,62,0,439,28
-Delaware,456,160,0,0,0,35,0
-Florida,9060,3447,26,247,0,4833,50
-Georgia,4032,3039,0,130,0,509,15
-Hawaii,23,0,0,162,71,16,29
-Idaho,763,0,15,11,0,109,0
-Illinois,6134,2274,11,430,0,2026,77
-Indiana,4320,753,0,45,0,221,0
-Iowa,2372,89,22,43,11,100,0
-Kansas,1278,143,27,22,0,179,23
-Kentucky,1825,240,0,20,0,60,0
-Louisiana,2856,2495,19,36,0,159,0
-Maine,208,0,0,0,0,0,0
-Maryland,2141,1958,0,189,0,535,20
-Massachusetts,6856,731,0,263,0,596,97
-Michigan,4687,2618,22,77,0,235,110
-Minnesota,2818,224,41,135,0,124,12
-Mississippi,1863,1607,86,18,0,42,0
-Missouri,3450,668,10,45,0,99,16
-Montana,444,0,153,0,0,17,0
-Nebraska,948,55,11,12,0,128,0
-Nevada,856,238,21,243,17,584,0
-New Hampshire,392,0,0,0,0,13,91
-New Jersey,8228,2686,11,917,0,3155,55
-New Mexico,318,13,523,0,0,355,0
-New York,8204,1884,24,475,0,1857,86
-New York City,5584,6086,0,2076,0,6513,869
-North Carolina,1770,967,36,53,0,281,0
-North Dakota,634,0,61,0,0,15,0
-Ohio,5156,1019,0,67,0,122,14
-Oklahoma,1564,133,228,32,16,145,52
-Oregon,581,17,18,26,0,114,0
-Pennsylvania,7782,1860,0,225,0,563,29
-Rhode Island,1029,75,0,23,0,108,0
-South Carolina,2462,1584,12,27,0,119,0
-South Dakota,668,0,98,13,0,10,0
-Tennessee,3147,936,0,32,0,169,0
-Texas,7437,2366,42,415,10,11595,30
-Utah,597,0,44,29,38,165,0
-Virginia,2448,1114,0,168,0,373,22
-Washington,1645,76,67,165,44,342,26
-West Virginia,405,18,0,0,0,0,0
-Wisconsin,2910,287,40,60,0,232,0
-Wyoming,140,0,24,0,0,17,0
