+++ server/api/ageSex.js
@@ -1,8 +1,63 @@
 const router = require('express').Router()
 const AgeSex = require('../db/models/age-and-sex')
 const Sequelize = require('sequelize')
+const {func} = require('prop-types')
 const Op = Sequelize.Op
 module.exports = router
+
+const states = {
+  Alabama: 'AL',
+  Alaska: 'AK',
+  Arizona: 'AZ',
+  Arkansas: 'AR',
+  California: 'CA',
+  Colorado: 'CO',
+  Connecticut: 'CT',
+  Delaware: 'DE',
+  'District Of Columbia': 'DC',
+  Florida: 'FL',
+  Georgia: 'GA',
+  Hawaii: 'HI',
+  Idaho: 'ID',
+  Illinois: 'IL',
+  Indiana: 'IN',
+  Iowa: 'IA',
+  Kansas: 'KS',
+  Kentucky: 'KY',
+  Louisiana: 'LA',
+  Maine: 'ME',
+  Maryland: 'MD',
+  Massachusetts: 'MA',
+  Michigan: 'MI',
+  Minnesota: 'MN',
+  Mississippi: 'MS',
+  Missouri: 'MO',
+  Montana: 'MT',
+  Nebraska: 'NE',
+  Nevada: 'NV',
+  'New Hampshire': 'NH',
+  'New Jersey': 'NJ',
+  'New Mexico': 'NM',
+  'New York': 'NY',
+  'North Carolina': 'NC',
+  'North Dakota': 'ND',
+  Ohio: 'OH',
+  Oklahoma: 'OK',
+  Oregon: 'OR',
+  Pennsylvania: 'PA',
+  'Rhode Island': 'RI',
+  'South Carolina': 'SC',
+  'South Dakota': 'SD',
+  Tennessee: 'TN',
+  Texas: 'TX',
+  Utah: 'UT',
+  Vermont: 'VT',
+  Virginia: 'VA',
+  Washington: 'WA',
+  'West Virginia': 'WV',
+  Wisconsin: 'WI',
+  Wyoming: 'WY'
+}
 
 router.get('/', async (req, res, next) => {
   try {
@@ -71,11 +126,18 @@ router.get('/gender', async (req, res, next) => {
         males: data[key].maleCount
       })
     }
-    let columns = arr.map(elem => {
-      return elem.state
+
+    const data2 = arr.map(elem => {
+      if (elem.state) {
+        elem.state = states[elem.state]
+        elem.total = +elem.females + +elem.males
+        elem.females = +elem.females
+        elem.males = +elem.males
+      }
+      return elem
     })
-    arr.push(columns)
-    res.json(arr)
+
+    res.json(data2)
   } catch (error) {
     next(error)
   }
